Artem(ekalka) issues fixed
