+++ src/constants/cfg.tsx
@@ -1,0 +1,15 @@
+import { ColumnsType } from 'antd/lib/table'; 
+import DataType from '../interfaces/pagination-data'; 
+
+export const columns: ColumnsType<DataType> = [ 
+    { 
+      title: 'Страна', 
+      dataIndex: 'country', 
+      key: 'country', 
+    }, 
+    { 
+      title: 'Название школы', 
+      dataIndex: 'name', 
+      key: 'name', 
+    }, 
+  ];
